requirements.txt
argparse
Pillow
tqdm

